Lift proxy routes
